refactor(api): extract mod for API responses
