+++ src/apis/mod.rs
@@ -1,4 +1,5 @@
 pub mod middlewares;
 pub mod resources;
+pub mod responses;
 pub mod routes;
 pub mod server;
